[#25]api 나오지않아서 틀만 완성
