+++ src/components/Calender.jsx
@@ -1,12 +1,19 @@
 import { useEffect } from "react";
 import { useState } from "react";
 import * as S from "../styles/index.style";
-const Calendarcom = ({ onDayClick }) => {
+import { setDate } from "../store/module/Date";
+import { useDispatch } from "react-redux";
+const Calendarcom = () => {
   const [currentDate, setCurrentDate] = useState(new Date());
 
   const Tdate = {
     Year: currentDate.getFullYear(),
     Month: currentDate.getMonth() + 1,
+  };
+  const dispatch = useDispatch();
+
+  const onDayClick = (year, month, day, lastday) => {
+    dispatch(setDate({ Date: { year, month, day, lastday } }));
   };
   useEffect(() => {
     onDayClick(null);
@@ -40,6 +47,7 @@ const Calendarcom = ({ onDayClick }) => {
     const backBlanks = Array(back == 7 ? 0 : back).fill(null);
     const days = Array.from({ length: totalDays }, (_, i) => i + 1);
     const allDays = [...blanks, ...days, ...backBlanks];
+    console.log(Math.max(...allDays));
     return (
       <S.calender.CalendarGrid>
         {["일", "월", "화", "수", "목", "금", "토"].map((day) => (
@@ -49,7 +57,15 @@ const Calendarcom = ({ onDayClick }) => {
           <S.calender.Day
             key={index}
             onClick={
-              day ? () => onDayClick(Tdate.Year, Tdate.Month, day) : () => {}
+              day
+                ? () =>
+                    onDayClick(
+                      Tdate.Year,
+                      Tdate.Month,
+                      day,
+                      Math.max(...allDays)
+                    )
+                : () => {}
             }
           >
             {<p>{day}</p>}
