+++ src/store/config/configure.js
@@ -1,10 +1,19 @@
 import { configureStore } from "@reduxjs/toolkit";
 import nav from "../module/Dashboard";
 import user from "../module/User";
+import team from "../module/Message";
+import Date from "../module/Date";
 import { combineReducers } from "redux";
 import storage from "redux-persist/lib/storage";
 import { persistReducer, persistStore } from "redux-persist";
-import { FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER } from 'redux-persist';
+import {
+  FLUSH,
+  REHYDRATE,
+  PAUSE,
+  PERSIST,
+  PURGE,
+  REGISTER,
+} from "redux-persist";
 
 const persistConfig = {
   key: "root",
@@ -15,6 +24,8 @@ const persistConfig = {
 const rootReducer = combineReducers({
   nav,
   user,
+  team,
+  Date,
 });
 
 const persistedReducer = persistReducer(persistConfig, rootReducer);
@@ -28,6 +39,5 @@ export const store = configureStore({
       },
     }),
 });
-
 
 export const persistor = persistStore(store);
